cover the case where systemd is managing Hubble in simple process tracking mode

+++ hubblestack/daemon.py
@@ -47,9 +47,11 @@ def run():
         sys.exit(0)
 
     if __opts__['daemonize']:
-        check_pidfile()
+        check_pidfile(kill_other=True)
         salt.utils.daemonize()
         create_pidfile()
+    else:
+        check_pidfile(kill_other=False)
 
     signal.signal(signal.SIGTERM, clean_up_process)
     signal.signal(signal.SIGINT, clean_up_process)
@@ -421,10 +423,15 @@ def parse_args():
                         help='Any arguments necessary for a single function run')
     return vars(parser.parse_args())
 
-def check_pidfile():
+def check_pidfile(kill_other=False):
     '''
     Check to see if there's already a pidfile. If so, check to see if the
     indicated process is alive and is Hubble.
+
+    kill_other
+        Default false, if set to true, attempt to kill detected running Hubble
+        processes; otherwise exit with an error.
+
     '''
     pidfile = __opts__['pidfile']
     if os.path.isfile(pidfile):
@@ -441,14 +448,18 @@ def check_pidfile():
                     with open("/proc/{pid}/cmdline".format(pid=xpid),'r') as f2:
                         cmdline = f2.readline().strip().strip('\x00').replace('\x00',' ')
                         if 'hubble' in cmdline:
-                            log.warn("process seems to still be alive and is hubble, attempting to shutdown")
-                            os.kill(int(xpid), signal.SIGTERM)
-                            time.sleep(1)
-                            if os.path.isdir("/proc/{pid}".format(pid=xpid)):
-                                log.error("failed to shutdown process successfully; abnormal program exit")
-                                exit(1)
+                            if kill_other:
+                                log.warn("process seems to still be alive and is hubble, attempting to shutdown")
+                                os.kill(int(xpid), signal.SIGTERM)
+                                time.sleep(1)
+                                if os.path.isdir("/proc/{pid}".format(pid=xpid)):
+                                    log.error("failed to shutdown process successfully; abnormal program exit")
+                                    exit(1)
+                                else:
+                                    log.info("shutdown seems to have succeeded, proceeding with startup")
                             else:
-                                log.info("shutdown seems to have succeeded, proceeding with startup")
+                                log.error("refusing to run while another hubble instance is running")
+                                exit(1)
                         else:
                             log.info("process does not appear to be hubble, ignoring")
